requirements.txt
langchain==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.6
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
pandas
numpy